Move `clean_links` to embed/utils.py
This function will be shared between APIv2 and APIv3.
